test: split out the tests into modules
