+++ test_overwrite.py
@@ -1,0 +1,6 @@
+def main():
+    print("Weclome to gen-cli!")
+
+
+if __name__ == "__main__":
+    main()
